requirements.txt
asgiref==3.2.3
certifi==2021.10.8
charset-normalizer==2.0.11
defusedxml==0.7.1
Django==3.0.3
django-filter==21.1
djangorestframework==3.13.1
httpie==3.0.2
idna==3.3
importlib-metadata==4.10.1
Markdown==3.3.6
multidict==6.0.2
psycopg2-binary==2.8.6
Pygments==2.11.2
PySocks==1.7.1
pytz==2019.3
requests==2.27.1
requests-toolbelt==0.9.1
sqlparse==0.3.0
urllib3==1.26.8
zipp==3.7.0
